Two Mean-Reversal Strategies
I compared the performance of two mean-reversal strategies. One strategy is to use the volatility-based formula to calculate the position of the market portfolio. Another one is a return-based asset selection portfolio. Buy the bottom 20 percent quantile of stocks and sell the top 20 percent quantile of stocks among all S&P500 stocks. Hold those stocks until they are not within the 40 percent of the quantile.