Compare Two Types of Mean-Reversal Strategy
I developed two types of mean-reversal strategy. One is based on a mathematic formula that the difference between weekly volatility and a week of daily volatility is negative because mean-reversal makes daily volatility larger than weekly volatility. One is based on the market performance of a range of stocks, so long the highest negative return stocks and short the highest postive return stocks. The outcome indicated that the second portfolio performed best.